+++ src/tools/frontend.py
@@ -1,3 +1,4 @@
+import os
 import random
 import sys
 from pathlib import Path
@@ -7,8 +8,8 @@ import streamlit as st
 base_path = Path(__file__).parent.parent.parent
 sys.path.append(base_path.as_posix())
 
+from src.config import TARGET_PATH
 from src.tools.search_by_keyword import search_by_keyword
-from src.tools.clear_output_folder import clear_output_folder
 
 
 def main():
@@ -52,6 +53,15 @@ def main():
                 images_sample = random.sample(images, min(len(images), 20))
                 ex.image(images_sample, width=200)
 
+def clear_output_folder():
+    if not os.path.isdir(TARGET_PATH):
+        return
+    for root, dirs, files in os.walk(TARGET_PATH, topdown=False):
+        for name in files:
+            os.remove(os.path.join(root, name))
+        for name in dirs:
+            os.rmdir(os.path.join(root, name))
+    print(f"Output folder cleared.")
 
 if __name__ == "__main__":
     main()
